BYT: Install ufo exported header files in the same way as ufosrc
BZ: 140045

Use semi-deprecated but very commonly used Android build system
variables LOCAL_COPY_HEADERS and LOCAL_COPY_HEADERS_TO to install
header files in the same way as is done by ufosrc.  This allows other
modules to find the files in a common place.

This change is in response to a developer request.  It affects only the
build.

The ufosrc project exports several header files for use by other
projects using Android build system variables LOCAL_COPY_HEADERS and
LOCAL_COPY_HEADERS_TO.  This results in the header files being copied
to the target/product output directory (not to the target itself)
under directory obj/include.

Conversely, the ufo pre-built binary project merely included the header
files under an "include" directory from which they could be referenced.

This results in a build-time difference between using the pre-built
binaries versus building with ufo source in the tree.  This commit
removes that difference so that the header files are available in a
consistent way.

Variable LOCAL_COPY_HEADERS contains the name of each header file to
be copied out, and LOCAL_COPY_HEADERS_TO contains the directory name
into which the files are to be copied.

Change-Id: I1a77f045c9439013f7d7f8847b93282397551d63
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>