Disable GLES3.1 APIs due to legal contraint
BZ: 175544

By default, ufo driver has the GLES3.1 API support, which is POR for BYT but not
for PV. Per NDA contraint, we cannot share/discuss information about ES3.1 with
customers that are not Khronos members. This limitation will belifted in ww12.1.
But before that, we will have to turn GLES3.1 API off.

Change-Id: I5c6769b4e4995751979cb48d501cb6bdfb36c8d6
Signed-off-by: Yunwei Zhang <yunwei.zhang@intel.com>